requirements.txt
anndata==0.10.7
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
scanpy==1.10.1
scipy==1.13.1
seaborn==0.13.2
scikit-learn==1.5.0
squidpy==1.5.0
statsmodels==0.14.2
